defined functions before exporting in dbcontroller.js, renamed db folder to controllers and updated paths
