+++ server/routes/index.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const path = require('path')
-const db = require('../db/dbcontroller.js')
+const db = require('../controllers/dbcontroller.js')
 
 // Serve sign in page
 router.get('/', (req, res) => {
@@ -12,7 +12,6 @@ router.post('/signin', db.signInUser)
 
 router.post('/signup', db.registerUser)
 
-// handle logout
 router.get('/logout', (req, res, next) => {
 	if (req.session) {
 		req.session.destroy(error => {
